feat: refactor logic
